requirements.txt
torch==1.10.0
torchvision==0.11.1
kornia
matplotlib
Pillow
scikit-learn
scipy
seaborn
tensorboardX
timm
tqdm